requirements.txt
crewai>=0.11.0
streamlit>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
duckduckgo-search>=4.1.1
pandas>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0 